requirements.txt
# Core AI framework - requires 1.0+ for load_mcp_servers function
pydantic-ai-slim[anthropic,mcp]>=1.0

# API client - will be constrained by pydantic-ai-slim
anthropic>=0.75.0

# Utilities
python-dotenv~=1.0.1
langsmith~=0.1.147

# Airbyte connectors
airbyte-agent-gong
airbyte-agent-hubspot

# UI framework
gradio~=4.44.1
